requirements.txt
uvicorn
fastapi
loguru
pydantic
weasyprint
pillow
pathlib
pdfminer-six
pymupdf
patool